move logger to right place
