+++ neptulon/auth.go
@@ -1,0 +1,3 @@
+package neptulon
+
+type Auth struct{}
