+++ neptulon/jsonrpc/jsonrpc.go
@@ -1,7 +1,11 @@
 // Package jsonrpc implements JSON-RPC 2.0 protocol for Neptulon framework.
 package jsonrpc
 
-import "github.com/nbusy/devastator/neptulon"
+import (
+	"encoding/json"
+
+	"github.com/nbusy/devastator/neptulon"
+)
 
 // App is a Neptulon JSON-RPC app.
 type App struct {
@@ -9,5 +13,24 @@ type App struct {
 }
 
 // NewApp creates a Neptulon JSON-RPC app.
-func NewApp() {
+func NewApp(n *neptulon.App) *App {
+	a := App{}
+	n.Middleware(a.handler)
+	return &a
+}
+
+// Middleware registers a new middleware to handle incoming messages.
+func (a *App) Middleware(middleware func(conn *neptulon.Conn, session *neptulon.Session, msg *Message)) {
+	a.middleware = append(a.middleware, middleware)
+}
+
+func (a *App) handler(conn *neptulon.Conn, session *neptulon.Session, msg []byte) {
+	var m Message
+	if err := json.Unmarshal(msg, &m); err != nil {
+		return
+	}
+
+	for _, mid := range a.middleware {
+		mid(conn, session, &m)
+	}
 }
