+++ neptulon/jsonrpc/router.go
@@ -1,21 +1,17 @@
 package jsonrpc
 
-import (
-	"encoding/json"
-
-	"github.com/nbusy/devastator/neptulon"
-)
-
-// NewRouter creates a JSON-RPC 2.0 router instance and registers it with the Neptulon app.
-func NewRouter(app *neptulon.App) *Router {
-	r := Router{}
-	app.Middleware(r.middleware)
-	return &r
-}
+import "github.com/nbusy/devastator/neptulon"
 
 // Router is a JSON-RPC request routing middleware.
 type Router struct {
 	routes map[string]func(conn *neptulon.Conn, session *neptulon.Session, msg *Message)
+}
+
+// NewRouter creates a JSON-RPC router instance and registers it with the Neptulon JSON-RPC app.
+func NewRouter(app *App) *Router {
+	r := Router{}
+	app.Middleware(r.middleware)
+	return &r
 }
 
 // Route adds a new route registry.
@@ -23,11 +19,6 @@ func (r *Router) Route(route string, handler func(conn *neptulon.Conn, session *
 	r.routes[route] = handler
 }
 
-func (r *Router) middleware(conn *neptulon.Conn, session *neptulon.Session, msg []byte) {
-	var m Message
-	if err := json.Unmarshal(msg, &m); err != nil {
-		return
-	}
-
-	r.routes[m.Method](conn, session, &m)
+func (r *Router) middleware(conn *neptulon.Conn, session *neptulon.Session, msg *Message) {
+	r.routes[msg.Method](conn, session, msg)
 }
