+++ neptulon/logger.go
@@ -1,4 +1,4 @@
-package devastator
+package neptulon
 
 // Logger provides low level request logging, performance metrics, and other metrics data.
 type Logger struct{}
