--- neptulon/middleware.go
@@ -1,5 +1,0 @@
-package neptulon
-
-// Middleware interface to be implemented by any middleware type.
-type Middleware interface {
-}
