+++ server.go
@@ -32,7 +32,7 @@ func NewServer(cert, privKey []byte, laddr string, debug bool) (*Server, error) 
 	}
 
 	// n.Middleware() // json rpc protocol
-	// p.Middleware() // public json rpc routes
+	// p.Middleware("auth.google") // public json rpc routes
 	// p.Middleware() // cert auth
 	// p.Middleware() // private json rpc routes
